+++ src/collective/taxonomy/jsonimpl.py
@@ -1,17 +1,20 @@
-from lxml import etree
-import os
+# -*- coding: utf-8 -*-
 import json
+import os
+
+from lxml import etree
 
 from BTrees.OOBTree import OOBTree
 from Products.Five.browser import BrowserView
 from plone import api
-from zope.component import queryMultiAdapter, queryUtility
+from zope.component import queryMultiAdapter
+from zope.component import queryUtility
 from zope.i18n import translate
 
 from collective.taxonomy import PATH_SEPARATOR
 from collective.taxonomy.i18n import CollectiveTaxonomyMessageFactory as _
-from collective.taxonomy.interfaces import get_lang_code
 from collective.taxonomy.interfaces import ITaxonomy
+from collective.taxonomy.interfaces import get_lang_code
 from collective.taxonomy.vdex import TreeExport
 
 
@@ -24,7 +27,7 @@ class EditTaxonomyData(TreeExport, BrowserView):
         utility_name = request.get('taxonomy', '')
         taxonomy = queryUtility(ITaxonomy, name=utility_name)
         if not taxonomy:
-            raise ValueError('Taxonomy `%s` could not be found.' % utility_name)
+            raise ValueError('Taxonomy `%s` could not be found.' % utility_name)  # noqa: E501
 
         self.taxonomy = taxonomy
 
@@ -113,9 +116,9 @@ class ImportJson(BrowserView):
                 data_for_taxonomy = self.generate_data_for_taxonomy(
                     tree['subnodes'], language)
 
-                taxonomy.data[language] = OOBTree()
-                for key, value in data_for_taxonomy:
-                    taxonomy.data[language][key] = value
+                # Update taxonomy and use 'clear' since we want to remain with
+                # just the items that were submitted by the user.
+                taxonomy.update(language, data_for_taxonomy, True)
 
             return json.dumps({
                 'status': 'info',