+++ src/collective/taxonomy/utility.py
@@ -5,6 +5,7 @@ from collective.taxonomy.interfaces import ITaxonomy
 from collective.taxonomy.interfaces import get_lang_code
 from collective.taxonomy.vocabulary import Vocabulary
 
+from BTrees.IOBTree import IOBTree
 from BTrees.OOBTree import OOBTree
 from OFS.SimpleItem import SimpleItem
 
@@ -17,6 +18,7 @@ from plone.dexterity.interfaces import IDexterityFTI
 from plone.memoize import ram
 
 from zope.interface import implementer
+from zope.globalrequest import getRequest
 from zope.lifecycleevent import modified
 
 import generated
@@ -24,17 +26,46 @@ import logging
 
 from copy import copy
 
-from collective.taxonomy import PATH_SEPARATOR
+from collective.taxonomy import (
+    LEGACY_PATH_SEPARATOR,
+    PATH_SEPARATOR,
+    PRETTY_PATH_SEPARATOR,
+    NODE,
+)
+
+try:
+    from plone.protect.auto import safeWrite
+except ImportError:
+    # plone.protect < 3.x compatibility
+    def safeWrite(obj, request):
+        pass
 
 
 logger = logging.getLogger("collective.taxonomy")
 
 
+def pop_value(d, compare_value, default=None):
+    for key, value in d.items():
+        if compare_value == value:
+            break
+    else:
+        return default
+
+    return d.pop(key)
+
+
 @implementer(ITaxonomy)
 class Taxonomy(SimpleItem):
+    order = None
+    count = None
+    version = None
 
     def __init__(self, name, title, default_language):
         self.data = PersistentDict()
+        self.order = PersistentDict()
+        self.count = PersistentDict()
+        self.version = PersistentDict()
+
         self.name = name
         self.title = title
         self.default_language = default_language
@@ -44,17 +75,12 @@ class Taxonomy(SimpleItem):
         return api.portal.get().getSiteManager()
 
     def __call__(self, context):
-
         if not self.data:
-            return Vocabulary(self.name, {}, {})
+            return Vocabulary(self.name, {}, {}, {}, 2)
 
         request = getattr(context, "REQUEST", None)
-
-        current_language = self.getCurrentLanguage(request)
-        data = self.data[current_language]
-        inverted_data = self.inverted_data[current_language]
-
-        return Vocabulary(self.name, data, inverted_data)
+        language = self.getCurrentLanguage(request)
+        return self.makeVocabulary(language)
 
     @property
     @ram.cache(lambda method, self: (self.name, self.data._p_mtime))
@@ -75,6 +101,14 @@ class Taxonomy(SimpleItem):
     def getVocabularyName(self):
         return 'collective.taxonomy.' + self.getShortName()
 
+    def makeVocabulary(self, language):
+        self._fixup()
+        data = self.data.get(language, {})
+        order = self.order.get(language)
+        version = self.version.get(language, 1)
+        inverted_data = self.inverted_data.get(language, {})
+        return Vocabulary(self.name, data, inverted_data, order, version)
+
     def getCurrentLanguage(self, request):
         language = get_lang_code()
         if language in self.data:
@@ -84,6 +118,23 @@ class Taxonomy(SimpleItem):
         else:
             # our best guess!
             return self.data.keys()[0]
+
+    def getLanguages(self):
+        return tuple(self.data)
+
+    def iterLanguage(self, language=None):
+        if language is None:
+            language = self.default_language
+
+        vocabulary = self.makeVocabulary(language)
+
+        for path, identifier in vocabulary.iterEntries():
+            parent_path = path.rsplit(PATH_SEPARATOR, 1)[0]
+            if parent_path:
+                parent = vocabulary.getTermByValue(parent_path)
+            else:
+                parent = None
+            yield path, identifier, parent
 
     def registerBehavior(self, **kwargs):
         new_args = copy(kwargs)
@@ -118,7 +169,11 @@ class Taxonomy(SimpleItem):
         if utility:
             utility.deactivateSearchable()
             utility.activateSearchable()
-            utility.title = kwargs['field_title']
+            if 'field_title' in kwargs:
+                utility.title = kwargs.pop('field_title')
+
+            for k, v in kwargs.iteritems():
+                setattr(utility, k, v)
 
         delattr(generated, short_name)
 
@@ -144,11 +199,99 @@ class Taxonomy(SimpleItem):
     def clean(self):
         self.data.clear()
 
-    def add(self, language, identifier, path):
-        if language not in self.data:
-            self.data[language] = OOBTree()
+    def add(self, language, value, key):
+        self._fixup()
+        tree = self.data.get(language)
+        if tree is None:
+            tree = self.data[language] = OOBTree()
+        else:
+            # Make sure we update the modification time.
+            self.data[language] = tree
 
-        self.data[language][path] = identifier
+        update = key in tree
+        tree[key] = value
+
+        order = self.order.get(language)
+        if order is None:
+            order = self.order[language] = IOBTree()
+            count = self.count[language] = 0
+        else:
+            if update:
+                pop_value(tree, key)
+
+            count = self.count[language] + 1
+
+        self.count[language] = count
+        order[count] = key
+
+    def update(self, language, items, clear=False):
+        self._fixup()
+
+        tree = self.data.setdefault(language, OOBTree())
+        if clear:
+            tree.clear()
+
+        # A new tree always uses the newest version.
+        if not tree:
+            version = self.version[language] = 2
+        else:
+            version = self.version.get(language, 1)
+
+        order = self.order.setdefault(language, IOBTree())
+        count = self.count.get(language, 0)
+
+        if clear:
+            order.clear()
+            count = 0
+
+        # Always migrate to newest version.
+        if version == 1:
+            def fix(path):
+                return path.replace(LEGACY_PATH_SEPARATOR, PATH_SEPARATOR)
+
+            for i in list(order):
+                path = order[i]
+                order[i] = fix(path)
+
+            for path in list(tree):
+                value = tree.pop(path)
+                tree[fix(path)] = value
+
+            version = self.version[language] = 2
+            logger.info(
+                "Taxonomy '%s' upgraded to version %d for language '%s'." % (
+                    self.name, version, language
+                )
+            )
+
+        # Make sure we update the modification time.
+        self.data[language] = tree
+
+        # The following structure is used to expunge updated entries.
+        inv = {}
+        if not clear:
+            for i, key in order.items():
+                inv[key] = i
+
+        seen = set()
+        for key, value in items:
+            if key in seen:
+                logger.warning("Duplicate key entry: %r" % (key, ))
+
+            seen.add(key)
+            update = key in tree
+            tree[key] = value
+            order[count] = key
+            count += 1
+
+            # If we're updating, then we have to pop out the old ordering
+            # information in order to maintain relative ordering of new items.
+            if update:
+                i = inv.get(key)
+                if i is not None:
+                    del order[i]
+
+        self.count[language] = count
 
     def translate(self, msgid, mapping=None, context=None,
                   target_language=None, default=None):
@@ -162,7 +305,28 @@ class Taxonomy(SimpleItem):
         if msgid not in self.inverted_data[target_language]:
             return ''
 
+        if self.version is not None and self.version.get(target_language) != 2:
+            path_sep = LEGACY_PATH_SEPARATOR
+        else:
+            path_sep = PATH_SEPARATOR
+
         path = self.inverted_data[target_language][msgid]
-        pretty_path = path[1:].replace(PATH_SEPARATOR, u' » ')
+        pretty_path = path[1:].replace(path_sep, PRETTY_PATH_SEPARATOR)
+
+        if mapping is not None and mapping.get(NODE):
+            pretty_path = pretty_path.rsplit(PRETTY_PATH_SEPARATOR, 1)[-1]
 
         return pretty_path
+
+    def _fixup(self):
+        # due to compatibility reasons this method fixes data structure
+        # for old Taxonomy instances.
+        # XXX: remove this in version 2.0 to prevent write on read
+        if self.order is None:
+            safeWrite(self, getRequest())
+            self.order = PersistentDict()
+            self.count = PersistentDict()
+
+        if self.version is None:
+            safeWrite(self, getRequest())
+            self.version = PersistentDict()